LCA Forest update
Previously the LCA queries only worked on a single tree. Now they work on a forest. Independent trees within the forest are handled in parallel. If vertices in a query are not in the same tree, then gives nullptr as the LCA, because there is no shared ancestor.

Not debugged yet.